move insert values's logic from RewriteEngine to InsertValuesTokenGenerator
